rename module to `on_disk_inverted_index`
