+++ lib/segment/src/index/field_index/full_text_index/inverted_index/immutable_inverted_index.rs
@@ -10,8 +10,8 @@ use itertools::Either;
 use posting_list::{PostingBuilder, PostingList, PostingListView, PostingValue};
 
 use super::immutable_postings_enum::ImmutablePostings;
-use super::mmap_inverted_index::MmapInvertedIndex;
-use super::mmap_inverted_index::mmap_postings_enum::OnDiskPostingsEnum;
+use super::on_disk_inverted_index::MmapInvertedIndex;
+use super::on_disk_inverted_index::on_disk_postings_enum::OnDiskPostingsEnum;
 use super::mutable_inverted_index::MutableInvertedIndex;
 use super::positions::Positions;
 use super::postings_iterator::{
