+++ lib/segment/src/index/field_index/full_text_index/inverted_index/on_disk_inverted_index/on_disk_postings.rs
@@ -11,8 +11,8 @@ use zerocopy::FromBytes;
 
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::full_text_index::inverted_index::TokenId;
-use crate::index::field_index::full_text_index::inverted_index::mmap_inverted_index::raw_posting_list::RawPostingList;
-use crate::index::field_index::full_text_index::inverted_index::mmap_inverted_index::types::{
+use crate::index::field_index::full_text_index::inverted_index::on_disk_inverted_index::raw_posting_list::RawPostingList;
+use crate::index::field_index::full_text_index::inverted_index::on_disk_inverted_index::types::{
     PostingListHeader, PostingsHeader, ZerocopyPostingValue,
 };
