+++ lib/segment/src/index/field_index/full_text_index/inverted_index/on_disk_inverted_index/on_disk_postings_enum.rs
@@ -5,7 +5,7 @@ use common::universal_io::UniversalRead;
 use super::super::positions::Positions;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::full_text_index::inverted_index::TokenId;
-use crate::index::field_index::full_text_index::inverted_index::mmap_inverted_index::uio_postings::OnDiskPostings;
+use crate::index::field_index::full_text_index::inverted_index::on_disk_inverted_index::on_disk_postings::OnDiskPostings;
 
 pub enum OnDiskPostingsEnum<S: UniversalRead> {
     Ids(OnDiskPostings<(), S>),
