+++ lib/segment/src/index/field_index/full_text_index/on_disk_text_index/mod.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 use common::bitvec::BitVec;
 use common::universal_io::{MmapFile, UniversalRead};
 
-use super::inverted_index::mmap_inverted_index::MmapInvertedIndex;
+use super::inverted_index::on_disk_inverted_index::MmapInvertedIndex;
 use super::inverted_index::mutable_inverted_index::MutableInvertedIndex;
 use super::tokenizers::Tokenizer;
 use crate::data_types::index::TextIndexParams;
